requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.28.0
python-dotenv==1.0.0
httpx==0.25.0
requests
